+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v2.5.4 (FUTURE)
 
 ## Bug Fixes
 
+* [#2779](https://github.com/digitalocean/netbox/issues/2779) - Include "none" option when filter IP addresses by role
 * [#2783](https://github.com/digitalocean/netbox/issues/2783) - Fix AttributeError exception when attempting to delete region(s)
 
 ---